chore: add gitignore
