+++ README.md
@@ -22,6 +22,7 @@
   - 支持自定义 API 端点
   - 可配置 AI 模型参数
   - 灵活的温度和令牌设置
+  - 配置保存在浏览器本地，无需部署时预先设置
 
 - 🎨 **优雅的用户界面**：
   - 响应式设计
@@ -48,17 +49,16 @@ cd English_analyze
 ```bash
 npm install
 ```
-
-3. 配置API
-- 编辑 `config/default.json`，填入您的 OpenAI API 端点、密钥等配置
-
-4. 启动应用
+ 
+3. 启动应用
 ```bash
 npm start
 ```
 
-5. 访问应用
+4. 访问应用
 打开浏览器访问 `http://localhost:8080`
+
+5. 在页面右上角的“设置”中配置您的 API 端点与密钥
 
 
 ## 📖 使用说明
