--- config/default.json
@@ -1,12 +1,0 @@
-{
-  "server": {
-    "port": 8080
-  },
-  "ai": {
-    "apiEndpoint": "https://api.openai.com/v1",
-    "apiKey": "your-api-key",
-    "model": "gpt-4o",
-    "temperature": 0.7,
-    "maxTokens": 2000
-  }
-}
