+++ package.json
@@ -16,12 +16,9 @@
   "author": "",
   "license": "MIT",
   "dependencies": {
-    "axios": "^1.6.2",
-    "config": "^3.3.9",
-    "express": "^4.18.2",
-    "openai": "^4.20.0"
+    "express": "^4.18.2"
   },
   "devDependencies": {
     "nodemon": "^3.0.1"
   }
-} 
+}
