+++ public/js/main.js
@@ -5,7 +5,6 @@ document.addEventListener('DOMContentLoaded', function() {
   const wordMode = document.getElementById('wordMode');
   const loadingIndicator = document.getElementById('loadingIndicator');
   const resultContainer = document.getElementById('resultContainer');
-  const settingsForm = document.getElementById('settingsForm');
   const saveSettingsBtn = document.getElementById('saveSettings');
 
   // 加载配置
@@ -18,54 +17,108 @@ document.addEventListener('DOMContentLoaded', function() {
   });
   saveSettingsBtn.addEventListener('click', saveConfig);
 
-  // 加载配置
-  async function loadConfig() {
-    try {
-      const response = await fetch('/api/config');
-      if (!response.ok) throw new Error('获取配置失败');
-      
-      const config = await response.json();
-      
-      // 填充表单
-      document.getElementById('apiEndpoint').value = config.apiEndpoint || '';
-      document.getElementById('model').value = config.model || '';
-      document.getElementById('temperature').value = config.temperature || 0.7;
-      document.getElementById('maxTokens').value = config.maxTokens || 1000;
-    } catch (error) {
-      showToast('加载配置失败: ' + error.message, 'error');
+  // 提示模板
+  function wordPrompt(word) {
+    return `请以JSON格式提供英语单词"${word}"的详细分析，包括以下内容：
+1. 单词拼写和音标（英式和美式）
+2. 所有词性及对应的中文释义
+3. 常见用法和搭配
+4. 3个包含该单词的例句（英文+中文翻译）
+5. 词源和记忆技巧（必须使用中文解释）
+
+格式要求：
+{
+  "word": "单词",
+  "phonetics": {
+    "uk": "英式音标",
+    "us": "美式音标"
+  },
+  "definitions": [
+    {
+      "partOfSpeech": "词性",
+      "meanings": ["中文释义1", "中文释义2"]
     }
+  ],
+  "usages": ["常见搭配1", "常见搭配2"],
+  "examples": [
+    {
+      "english": "英文例句1",
+      "chinese": "中文翻译1"
+    }
+  ],
+  "etymology": "词源简介（请使用中文）",
+  "tips": "记忆技巧（请使用中文）"
+}
+
+请确保JSON格式正确可被直接解析，并且词源和记忆技巧部分必须使用中文解释，不要使用英文。`;
+  }
+
+  function sentencePrompt(sentence) {
+    return `请严格按照JSON格式提供对英语句子"${sentence}"的详细解析。
+
+要求：
+1. 必须返回符合JSON规范的数据
+2. 不要包含任何额外的解释性文字
+3. 所有字符串必须使用双引号而不是单引号
+4. 使用中文进行解释分析
+
+格式要求：
+{
+  "sentence": "${sentence}",
+  "translation": "整体中文翻译",
+  "structure": {
+    "type": "句子类型",
+    "explanation": "结构解释"
+  },
+  "components": [
+    {
+      "text": "句子片段",
+      "role": "句法角色",
+      "explanation": "解释"
+    }
+  ],
+  "keyPhrases": [
+    {
+      "phrase": "短语或词汇",
+      "meaning": "含义",
+      "usage": "用法说明"
+    }
+  ],
+  "grammar": [
+    {
+      "aspect": "语法点",
+      "explanation": "详细解释"
+    }
+  ]
+}`;
+  }
+
+  // 加载配置
+  function loadConfig() {
+    const config = JSON.parse(localStorage.getItem('aiConfig') || '{}');
+    document.getElementById('apiEndpoint').value = config.apiEndpoint || 'https://api.openai.com/v1';
+    document.getElementById('apiKey').value = config.apiKey || '';
+    document.getElementById('model').value = config.model || 'gpt-4o';
+    document.getElementById('temperature').value = config.temperature ?? 0.7;
+    document.getElementById('maxTokens').value = config.maxTokens ?? 1000;
   }
 
   // 保存配置
-  async function saveConfig() {
-    try {
-      const config = {
-        apiEndpoint: document.getElementById('apiEndpoint').value,
-        apiKey: document.getElementById('apiKey').value,
-        model: document.getElementById('model').value,
-        temperature: parseFloat(document.getElementById('temperature').value),
-        maxTokens: parseInt(document.getElementById('maxTokens').value)
-      };
+  function saveConfig() {
+    const config = {
+      apiEndpoint: document.getElementById('apiEndpoint').value,
+      apiKey: document.getElementById('apiKey').value,
+      model: document.getElementById('model').value,
+      temperature: parseFloat(document.getElementById('temperature').value),
+      maxTokens: parseInt(document.getElementById('maxTokens').value)
+    };
 
-      const response = await fetch('/api/config', {
-        method: 'POST',
-        headers: {
-          'Content-Type': 'application/json'
-        },
-        body: JSON.stringify(config)
-      });
+    localStorage.setItem('aiConfig', JSON.stringify(config));
+    showToast('配置已保存', 'success');
 
-      if (!response.ok) throw new Error('保存配置失败');
-      
-      const result = await response.json();
-      showToast(result.message, 'success');
-      
-      // 关闭模态框
-      const modal = bootstrap.Modal.getInstance(document.getElementById('settingsModal'));
-      modal.hide();
-    } catch (error) {
-      showToast('保存配置失败: ' + error.message, 'error');
-    }
+    // 关闭模态框
+    const modal = bootstrap.Modal.getInstance(document.getElementById('settingsModal'));
+    modal.hide();
   }
 
   // 执行分析
@@ -77,32 +130,76 @@ document.addEventListener('DOMContentLoaded', function() {
     }
 
     const mode = wordMode.checked ? 'word' : 'sentence';
-    
+    const config = JSON.parse(localStorage.getItem('aiConfig') || '{}');
+    const { apiEndpoint, apiKey, model, temperature, maxTokens } = config;
+
+    if (!apiKey) {
+      showToast('请先在设置中配置API密钥', 'warning');
+      return;
+    }
+
+    const prompt = mode === 'word' ? wordPrompt(text) : sentencePrompt(text);
+
     // 显示加载指示器
     loadingIndicator.classList.remove('d-none');
     resultContainer.classList.add('d-none');
 
     try {
-      const response = await fetch('/api/analyze', {
+      const response = await fetch(`${apiEndpoint}/chat/completions`, {
         method: 'POST',
         headers: {
-          'Content-Type': 'application/json'
+          'Content-Type': 'application/json',
+          'Authorization': `Bearer ${apiKey}`
         },
-        body: JSON.stringify({ text, mode })
+        body: JSON.stringify({
+          model,
+          messages: [
+            {
+              role: 'system',
+              content: mode === 'word'
+                ? '你是一个专业的英语教师助手，擅长解析英语单词。请始终以JSON格式返回响应。'
+                : '你是一个专业的英语教师助手，擅长解析英语句子。请始终以JSON格式返回响应，不要包含任何额外的文字说明。'
+            },
+            {
+              role: 'user',
+              content: prompt
+            }
+          ],
+          temperature,
+          max_tokens: maxTokens
+        })
       });
 
       if (!response.ok) throw new Error('分析请求失败');
-      
+
       const data = await response.json();
-      
-      if (mode === 'word') {
-        renderWordResult(data.result);
-      } else {
-        renderSentenceResult(data.result);
+      if (!data?.choices?.[0]?.message?.content) {
+        throw new Error('API返回的数据格式不完整');
       }
 
-      resultContainer.classList.remove('d-none');
-      resultContainer.scrollIntoView({ behavior: 'smooth' });
+      let content = data.choices[0].message.content.trim();
+      try {
+        const jsonStart = content.indexOf('{');
+        const jsonEnd = content.lastIndexOf('}');
+        if (jsonStart === -1 || jsonEnd === -1) {
+          throw new Error('返回内容中未找到有效的JSON结构');
+        }
+
+        content = content.slice(jsonStart, jsonEnd + 1);
+        content = content.replace(/```json\s*|```\s*/g, '');
+        const result = JSON.parse(content);
+
+        if (mode === 'word') {
+          renderWordResult(result);
+        } else {
+          renderSentenceResult(result);
+        }
+
+        resultContainer.classList.remove('d-none');
+        resultContainer.scrollIntoView({ behavior: 'smooth' });
+      } catch (parseError) {
+        showToast('解析结果失败: ' + parseError.message, 'error');
+      }
     } catch (error) {
       showToast('分析失败: ' + error.message, 'error');
     } finally {
